Update React ErrorHandler component reference
The ErrorHandler component moved from the CyVerseReactComponent's util directory to within ui-lib
